pyproject.toml
[project]
name = "firewalld-ext"
version = "0.1.75"
description = "Open source threat feed auto blocker."
authors = [
  { name = "Sisyphus1813", email = "Sisyphus1813@protonmail.com" },
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "systemd-python>=235",
]

[project.scripts]
firewalld-ext = "main:main"

[tool.pyrefly]
mode = "strict"
